Start splitting out two zone noise model stuff
