+++ src/bloqade/cirq_utils/noise/__init__.py
@@ -8,10 +8,9 @@ from .model import (
     GeminiOneZoneNoiseModelConflictGraphMoves as GeminiOneZoneNoiseModelConflictGraphMoves,
 )
 from .utils import (
-    get_equivalent_swaps as get_equivalent_swaps,
-    get_two_zoned_noisy_circ as get_two_zoned_noisy_circ,
     transform_to_qasm_u_gates as transform_to_qasm_u_gates,
     optimize_circuit_to_cz_gate_set as optimize_circuit_to_cz_gate_set,
     transform_to_noisy_one_zone_circuit as transform_to_noisy_one_zone_circuit,
 )
 from .conflict_graph import OneZoneConflictGraph as OneZoneConflictGraph
+from .two_zone_model import get_two_zoned_noisy_circ as get_two_zoned_noisy_circ
